Slice 16: Decompose PresetForm and dedupe action dialog
- Extract PresetFieldsSection (name/description/category/theme).
- Extract PresetActionsEditor (in-memory action template list + dialog).
- Add shared ActionFormDialog component used by both PresetActionsEditor
  and ActionsEditor; ActionsEditor now imports it instead of inlining
  the same 50 lines of dialog markup.
- Add file-header docblocks to every new file; PresetForm drops from
  309 to 108 lines.
